requirements.txt
openai
langchain
python-dotenv
fastapi
unicorn
streamlit
request
